+++ oci-test/tests/oci_cloud_guard_target/test-turbot-query.sql
@@ -1,0 +1,3 @@
+select title, tenant_id
+from oci.oci_cloud_guard_target
+where id = '{{ output.resource_id.value }}';