+++ oci-test/tests/oci_cloud_guard_target/variables.tf
@@ -1,0 +1,56 @@
+variable "resource_name" {
+  type        = string
+  default     = "steampipetest20200125"
+  description = "Name of the resource used throughout the test."
+}
+
+variable "tenancy_ocid" {
+  type        = string
+  description = "OCI credentials profile used for the test. Default is to use the default profile."
+}
+
+variable "config_file_profile" {
+  type        = string
+  default     = "DEFAULT"
+  description = "OCI credentials profile used for the test. Default is to use the default profile."
+}
+
+variable "region" {
+  type        = string
+  default     = "ap-mumbai-1"
+  description = "OCI region used for the test. Does not work with default region in config, so must be defined here."
+}
+
+provider "oci" {
+  tenancy_ocid        = var.tenancy_ocid
+  config_file_profile = var.config_file_profile
+}
+
+locals {
+  path = "${path.cwd}/output.json"
+}
+
+resource "null_resource" "named_test_resource" {
+  provisioner "local-exec" {
+    command = "oci cloud-guard target list --compartment-id ${var.tenancy_ocid} --output json > ${local.path}"
+  }
+}
+
+data "local_file" "input" {
+  depends_on = [null_resource.named_test_resource]
+  filename   = local.path
+}
+
+output "resource_name" {
+  depends_on = [null_resource.named_test_resource]
+  value      = jsondecode(data.local_file.input.content).data.items[0].display-name
+}
+
+output "tenancy_ocid" {
+  value = var.tenancy_ocid
+}
+
+output "resource_id" {
+  depends_on = [null_resource.named_test_resource]
+  value      = jsondecode(data.local_file.input.content).data.items[0].id
+}