+++ oci-test/tests/oci_core_boot_volume_backup/test-list-query.sql
@@ -1,0 +1,3 @@
+select boot_volume_id, display_name, id, freeform_tags, lifecycle_state, size_in_gbs
+from oci.oci_core_boot_volume_backup
+where display_name = '{{ resourceName }}';