+++ oci-test/tests/oci_file_storage_snapshot/test-get-query.sql
@@ -1,0 +1,3 @@
+select id, name, lifecycle_state, is_clone_source, freeform_tags
+from oci.oci_file_storage_snapshot
+where id = '{{ output.resource_id.value }}';