+++ oci-test/tests/oci_kms_key/test-list-query.sql
@@ -1,0 +1,3 @@
+select name, id, lifecycle_state
+from oci.oci_kms_key
+where name = '{{ resourceName }}';