+++ oci/table_oci_cloud_guard_managed_list.go
@@ -110,7 +110,7 @@ func tableCloudGuardManagedList(_ context.Context) *plugin.Table {
 			},
 			{
 				Name:        "system_tags",
-				Description: "Tags added to instances by the service.",
+				Description: ColumnDescriptionSystemTags,
 				Type:        proto.ColumnType_JSON,
 			},